TST: move editable wheel tests to their own file
